wspace/joaquinrossi-esoteric: Remove deleted repo
When I first started tracking https://github.com/joaquin-rossi/esoteric,
it had interpreters for Whitespace and Brainfuck in Python, which I
archived [0]. It was later force-pushed and replaced with only a
Brainfuck interpreter in Rust. Now, the repo has been deleted, so there
is no need to continue tracking it here. The final known state here
matches the archive on SWH [1].

[0]: https://github.com/wspace/joaquinrossi-esoteric
[1]: https://archive.softwareheritage.org/browse/origin/directory/?origin_url=https://github.com/joaquin-rossi/esoteric